+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This release is under active development.  Actual release date TBD.
 New features:
 
 * Add a new option ``--set-server-host`` in scalyr-agent-2-config to set the server host from the commandline.
+* Add new instance variables ``_log_write_rate`` and ``_log_max_write_burst`` to ScalyrMonitor to allow monitor developers to override the rate limits imposed on their monitor's log.  See the comments in test_monitor.py for more details.
 
 Bug fixes: